Move second screen to own component
